modules/location-event/requirements.txt
grpcio == 1.41..1
grpcio-tools == 1.41.1
kafka-python == 2.0.2
grpcio-reflection==1.41.1